move outline_config to re_space_view
