+++ crates/viewer/re_viewer_context/src/gpu_bridge/mod.rs
@@ -2,7 +2,6 @@
 
 mod colormap;
 mod image_to_gpu;
-mod outline;
 mod re_renderer_callback;
 
 pub use colormap::{colormap_edit_or_view_ui, colormap_to_re_renderer};
@@ -10,7 +9,6 @@ pub use image_to_gpu::{
     image_data_range_heuristic, image_to_gpu, required_shader_decode,
     texture_creation_desc_from_color_image,
 };
-pub use outline::outline_config;
 pub use re_renderer_callback::new_renderer_callback;
 
 use crate::TensorStats;
